Isolated Pub Sub Manager
In addition, I have turned it into a plugin with the plugin structure.

Since there is no specific messaging in the pub sub manager for tasks, the is no separate logic file, it all is done in services which you will call via querying list of services and then calling the service.