run_device_viewer_pluggable_dropbot_service_demo.py contains a working examples of using a totally pluggable way of using dropbot services.
To test it, run the script and connect / disconnect a dropbot. Check the console logs.

One can add more functionality to the dropbot controller using mixin classes with more functions.
A message router plugin takes care of triuggering the right actors.

The paths for these actors, and the services offered by the mixins should be declared in a constants file within each plugin directory. It should also have the info on the subscriptions offered by the plugin.

This allows communication between plugins in a systematic fashion.

Run the