Provides message router plugin (#104)
* move around some files. Have plugins working seperately imported into example run scripts.

* Preferences for using default layout needs to be true. Only changed after the fact for furhter app intitializations to false.

* Import fixes

* fixup! Import fixes

* fixup! Import fixes

* folder name change

* got a working extensible dropbot service plugin. example script for demoing it is present as well.

* run_device_viewer_pluggable_dropbot_service_demo.py contains a working examples of using a totally pluggable way of using dropbot services.

To test it, run the script and connect / disconnect a dropbot. Check the console logs.

One can add more functionality to the dropbot controller using mixin classes with more functions.
A message router plugin takes care of triuggering the right actors.

The paths for these actors, and the services offered by the mixins should be declared in a constants file within each plugin directory. It should also have the info on the subscriptions offered by the plugin.

This allows communication between plugins in a systematic fashion.

Run the

* change a directory name in examples to make celar its just dummy examples.

---------

Co-authored-by: Vigneshwar Rajesh <vignesh.rajesh@outlook.com>